폴더구조 세팅 수정
data, ui 분리
