+++ app/src/main/java/com/example/woomansi/data/repository/RepositorySample.java
@@ -1,0 +1,5 @@
+package com.example.woomansi.data.repository;
+
+public class RepositorySample {
+
+}
