+++ app/src/main/java/com/example/woomansi/ui/adapter/RecyclerAdapterSample.java
@@ -1,0 +1,5 @@
+package com.example.woomansi.ui.adapter;
+
+public class RecyclerAdapterSample {
+
+}
