+++ app/src/main/java/com/example/woomansi/ui/screen/main/Main1Fragment.java
@@ -1,4 +1,4 @@
-package com.example.woomansi.ui.main;
+package com.example.woomansi.ui.screen.main;
 
 import android.os.Bundle;
