+++ app/src/main/java/com/example/woomansi/ui/viewmodel/MainViewModel.java
@@ -1,0 +1,7 @@
+package com.example.woomansi.ui.viewmodel;
+
+import androidx.lifecycle.ViewModel;
+
+public class MainViewModel extends ViewModel {
+
+}
