+++ app/src/main/res/layout/fragment_main3.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.Main3Fragment">
+    tools:context=".ui.screen.main.Main3Fragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
